Simplify: keep search_backend singleton in backends.py, drop netbox.search.backend
The prior commit extracted get_backend()/search_backend into a new
netbox.search.backend module to break the backends -> deferred -> jobs -> backends
cycle. That wasn't necessary: the actual cycle came from backends.py's own
module-level "from .deferred import ..." executing *before* search_backend got
defined further down the same file. Scoping that import to just the two methods
that need it (already done in the prior commit) is sufficient on its own --
deferred.py and jobs.py can import search_backend from backends.py at their own
module level without incident, since nothing forces deferred.py to load during
backends.py's own module-level execution anymore.

Keeping the singleton in backends.py avoids moving a symbol (get_backend) that at
least one plugin imports directly from netbox.search.backends, which would
otherwise force a version-matrix import shim for no benefit.
