+++ netbox/netbox/forms/search.py
@@ -4,7 +4,7 @@ from django import forms
 from django.utils.translation import gettext_lazy as _
 
 from netbox.search import LookupTypes
-from netbox.search.backend import search_backend
+from netbox.search.backends import search_backend
 
 LOOKUP_CHOICES = (
     ('', _('Partial match')),
